# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CastDevice": {
        "properties": {
          "address": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "port": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "port"
        ],
        "type": "object"
      },
      "CastSessionStatus": {
        "oneOf": [
          {
            "properties": {
              "status": {
                "enum": [
                  "idle"
                ],
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "status": {
                "enum": [
                  "connecting"
                ],
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "status": {
                "enum": [
                  "transcoding"
                ],
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "position_secs": {
                "format": "float",
                "type": [
                  "number",
                  "null"
                ]
              },
              "status": {
                "enum": [
                  "playing"
                ],
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "position_secs": {
                "format": "float",
                "type": [
                  "number",
                  "null"
                ]
              },
              "status": {
                "enum": [
                  "paused"
                ],
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "finished": {
                "type": "boolean"
              },
              "status": {
                "enum": [
                  "stopped"
                ],
                "type": "string"
              }
            },
            "required": [
              "finished",
              "status"
            ],
            "type": "object"
          },
          {
            "properties": {
              "message": {
                "type": "string"
              },
              "status": {
                "enum": [
                  "error"
                ],
                "type": "string"
              }
            },
            "required": [
              "message",
              "status"
            ],
            "type": "object"
          }
        ]
      },
      "CastStatusResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/CastSessionStatus"
          },
          "volume": {
            "format": "float",
            "type": "number"
          }
        },
        "required": [
          "session",
          "volume"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "PlayBody": {
        "properties": {
          "device_address": {
            "type": "string"
          },
          "device_port": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "track_id": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "device_address",
          "device_port",
          "track_id"
        ],
        "type": "object"
      },
...
      "SeekBody": {
        "properties": {
          "position_secs": {
            "format": "float",
            "type": "number"
          },
          "was_playing": {
            "type": "boolean"
          }
        },
        "required": [
          "position_secs",
          "was_playing"
        ],
        "type": "object"
      },
...
      "VolumeBody": {
        "properties": {
          "level": {
            "format": "float",
            "type": "number"
          }
        },
        "required": [
          "level"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "The server's configured `api_key`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Muorg music server HTTP API, shared by the desktop (Tauri), web and Android clients. Everything except `GET /api/health` and `GET /stream/{id}` requires an `Authorization: Bearer <api_key>` header.\n\n`GET /api/tracks` is paginated and defaults to 500 rows — a bare call returns the first page, not the catalog. Follow `X-Total-Count`.",
    "license": {
      "name": ""
    },
    "title": "Muorg Server API",
    "version": "2.42.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cast/devices": {
      "get": {
        "operationId": "get_devices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CastDevice"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Discovered devices"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Chromecast devices seen by the current mDNS sweep.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/discovery/start": {
      "post": {
        "operationId": "start_discovery",
        "responses": {
          "204": {
            "description": "Discovery started"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Begin mDNS discovery.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/discovery/stop": {
      "post": {
        "operationId": "stop_discovery",
        "responses": {
          "204": {
            "description": "Discovery stopped"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Stop mDNS discovery.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/pause": {
      "post": {
        "operationId": "pause",
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Pause the cast session.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/play": {
      "post": {
        "operationId": "play",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Start casting a track to a device.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/resume": {
      "post": {
        "operationId": "resume",
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Resume the cast session.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/seek": {
      "post": {
        "operationId": "seek",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Seek within the casting track.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/status": {
      "get": {
        "operationId": "get_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CastStatusResponse"
                }
              }
            },
            "description": "Session status"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Current cast session state and device volume.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/stop": {
      "post": {
        "operationId": "stop",
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Tear down the cast session.",
        "tags": [
          "Cast"
        ]
      }
    },
    "/api/cast/volume": {
      "post": {
        "operationId": "set_volume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Command accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Set device volume, 0.0–1.0.",
        "tags": [
          "Cast"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}